some review comments
