+++ orb_models/forcefield/atomic_system.py
@@ -8,7 +8,8 @@ from ase.calculators.singlepoint import SinglePointCalculator
 
 from orb_models.forcefield import featurization_utilities
 from orb_models.forcefield.base import AtomGraphs
-from orb_models.forcefield.property_definitions import PROPERTIES, PropertyDefinition
+from orb_models.forcefield.property_definitions import (PROPERTIES,
+                                                        PropertyDefinition)
 
 
 @dataclass
@@ -26,7 +27,6 @@ class SystemConfig:
     use_timestep_0: bool = True
 
 
-@dataclass
 class PropertyConfig:
     """Defines which properties should be calculated and stored on the AtomGraphs batch.
 
@@ -247,27 +247,3 @@ def ase_fix_atoms_to_tensor(atoms: ase.Atoms) -> Optional[torch.Tensor]:
             fixed_atoms = torch.zeros((len(atoms)), dtype=torch.bool)
             fixed_atoms[constraint.index] = True
     return fixed_atoms
-
-
-def make_property_definitions_from_config(
-    config: Optional[Dict] = None,
-) -> PropertyConfig:
-    """Get PropertyConfig object from config."""
-    if config is None:
-        return PropertyConfig()
-    assert all(
-        key in ["node", "edge", "graph"] for key in config
-    ), "Only node, edge and graph properties are supported."
-
-    node_properties = edge_properties = graph_properties = None
-    if config.get("node"):
-        node_properties = [name for name in config["node"]]
-    if config.get("edge"):
-        edge_properties = [name for name in config["edge"]]
-    if config.get("graph"):
-        graph_properties = [name for name in config["graph"]]
-    return PropertyConfig(
-        node_names=node_properties,
-        edge_names=edge_properties,
-        graph_names=graph_properties,
-    )
