+++ orb_models/forcefield/featurization_utilities.py
@@ -4,7 +4,6 @@ from typing import Callable, Optional, Tuple, Union
 
 import numpy as np
 import torch
-
 # TODO(Mark): Make pynanoflann optional
 from pynanoflann import KDTree as NanoKDTree
 from scipy.spatial import KDTree as SciKDTree
