+++ orb_models/forcefield/graph_regressor.py
@@ -7,7 +7,8 @@ import torch.nn as nn
 from orb_models.forcefield import base, segment_ops
 from orb_models.forcefield.gns import _KEY, MoleculeGNS
 from orb_models.forcefield.nn_util import build_mlp
-from orb_models.forcefield.property_definitions import PROPERTIES, PropertyDefinition
+from orb_models.forcefield.property_definitions import (PROPERTIES,
+                                                        PropertyDefinition)
 from orb_models.forcefield.reference_energies import REFERENCE_ENERGIES
 
 global HAS_WARNED_FOR_TF32_MATMUL
