+++ tests/test_base.py
@@ -1,6 +1,7 @@
 # type: ignore
 import pytest
 import torch
+
 from orb_models.forcefield import base
 from orb_models.forcefield.base import refeaturize_atomgraphs
